ligand_binding_gnn/requirements.txt
torch>=2.0
torchvision
numpy
polars
tqdm
pyyaml
hydra-core>=1.2
omegaconf
transformers
esm
scikit-learn
pandas
wandb
neptune-client
rdkit-pypi  # optional; only required for graph-mode
